cleanup bounty reviewers and add filter by role for all reviewers
