+++ lib/permissions/bounties/client.ts
@@ -1,4 +1,3 @@
 export * from './mapping';
 export * from './interfaces';
 export * from './availableBountyPermissions';
-export * from './mapBountyPermissions';
