+++ lib/permissions/bounties/index.ts
@@ -1,5 +1,4 @@
 export * from './interfaces';
 export * from './mapping';
-export * from './addBountyPermissionGroup';
 export * from './computeBountyPermissions';
 export * from './availableBountyPermissions';
