--- lib/permissions/validateAssigneeGroup.ts
@@ -1,7 +1,0 @@
-import type { AssignablePermissionGroupsWithPublic } from './interfaces';
-
-const assignableGroups: Exclude<AssignablePermissionGroupsWithPublic, 'any'>[] = ['user', 'role', 'space', 'public'];
-
-export function assigneeGroupIsValid(input: any): boolean {
-  return assignableGroups.indexOf(input) >= 0;
-}
