+++ lib/rewards/__tests__/setRewardUsers.spec.ts
@@ -1,4 +1,3 @@
-import type { TargetPermissionGroup } from '@charmverse/core/permissions';
 import type { Space, Role, User } from '@charmverse/core/prisma-client';
 import { testUtilsMembers, testUtilsUser } from '@charmverse/core/test';
 
@@ -48,10 +47,7 @@ describe('setRewardUsers', () => {
 
   it('should successfully set reviewers for a reward', async () => {
     const reviewerRole = await testUtilsMembers.generateRole({ createdBy: user.id, spaceId: space.id });
-    const reviewers: TargetPermissionGroup<'user' | 'role'>[] = [
-      { group: 'user', id: user.id },
-      { group: 'role', id: reviewerRole.id }
-    ];
+    const reviewers = [{ userId: user.id }, { roleId: reviewerRole.id }];
 
     const result = await setRewardUsers({
       rewardId: reward.id,
@@ -64,7 +60,7 @@ describe('setRewardUsers', () => {
   });
 
   it('should successfully set both allowed submitter roles and reviewers for a reward', async () => {
-    const reviewers: TargetPermissionGroup<'user'>[] = [{ group: 'user', id: user.id }];
+    const reviewers = [{ userId: user.id }];
     const result = await setRewardUsers({
       rewardId: reward.id,
       users: {
@@ -90,7 +86,7 @@ describe('setRewardUsers', () => {
       setRewardUsers({
         rewardId: reward.id,
         users: {
-          reviewers: [{ group: 'role', id: 'invalid-id' }]
+          reviewers: [{ roleId: 'invalid-id' }]
         }
       })
     ).rejects.toThrow(Error); // Modify this to a more specific error if needed.
